+++ instrumentation/google.golang.org/grpc/otelgrpc/filters/interceptor/filters.go
@@ -1,13 +1,11 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package filters // import "go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc/filters"
+package interceptor // import "go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc/filters/interceptor"
 
 import (
 	"path"
 	"strings"
-
-	"google.golang.org/grpc/stats"
 
 	"go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc"
 )
@@ -21,8 +19,20 @@ type gRPCPath struct {
 // and returns as gRPCPath object that has divided service and method names
 // https://github.com/grpc/grpc/blob/master/doc/PROTOCOL-HTTP2.md
 // If name is not FullMethod, returned gRPCPath has empty service field.
-func splitFullMethod(i *stats.RPCTagInfo) gRPCPath {
-	s, m := path.Split(i.FullMethodName)
+func splitFullMethod(i *otelgrpc.InterceptorInfo) gRPCPath {
+	var name string
+	switch i.Type {
+	case otelgrpc.UnaryServer:
+		name = i.UnaryServerInfo.FullMethod
+	case otelgrpc.StreamServer:
+		name = i.StreamServerInfo.FullMethod
+	case otelgrpc.UnaryClient, otelgrpc.StreamClient:
+		name = i.Method
+	default:
+		name = i.Method
+	}
+
+	s, m := path.Split(name)
 	if s != "" {
 		s = path.Clean(s)
 		s = strings.TrimLeft(s, "/")
@@ -36,8 +46,10 @@ func splitFullMethod(i *stats.RPCTagInfo) gRPCPath {
 
 // Any takes a list of Filters and returns a Filter that
 // returns true if any Filter in the list returns true.
-func Any(fs ...otelgrpc.Filter) otelgrpc.Filter {
-	return func(i *stats.RPCTagInfo) bool {
+//
+// Deprecated: Use stats handler instead.
+func Any(fs ...otelgrpc.InterceptorFilter) otelgrpc.InterceptorFilter {
+	return func(i *otelgrpc.InterceptorInfo) bool {
 		for _, f := range fs {
 			if f(i) {
 				return true
@@ -49,8 +61,10 @@ func Any(fs ...otelgrpc.Filter) otelgrpc.Filter {
 
 // All takes a list of Filters and returns a Filter that
 // returns true only if all Filters in the list return true.
-func All(fs ...otelgrpc.Filter) otelgrpc.Filter {
-	return func(i *stats.RPCTagInfo) bool {
+//
+// Deprecated: Use stats handler instead.
+func All(fs ...otelgrpc.InterceptorFilter) otelgrpc.InterceptorFilter {
+	return func(i *otelgrpc.InterceptorInfo) bool {
 		for _, f := range fs {
 			if !f(i) {
 				return false
@@ -62,21 +76,27 @@ func All(fs ...otelgrpc.Filter) otelgrpc.Filter {
 
 // None takes a list of Filters and returns a Filter that returns
 // true only if none of the Filters in the list return true.
-func None(fs ...otelgrpc.Filter) otelgrpc.Filter {
+//
+// Deprecated: Use stats handler instead.
+func None(fs ...otelgrpc.InterceptorFilter) otelgrpc.InterceptorFilter {
 	return Not(Any(fs...))
 }
 
 // Not provides a convenience mechanism for inverting a Filter.
-func Not(f otelgrpc.Filter) otelgrpc.Filter {
-	return func(i *stats.RPCTagInfo) bool {
+//
+// Deprecated: Use stats handler instead.
+func Not(f otelgrpc.InterceptorFilter) otelgrpc.InterceptorFilter {
+	return func(i *otelgrpc.InterceptorInfo) bool {
 		return !f(i)
 	}
 }
 
 // MethodName returns a Filter that returns true if the request's
 // method name matches the provided string n.
-func MethodName(n string) otelgrpc.Filter {
-	return func(i *stats.RPCTagInfo) bool {
+//
+// Deprecated: Use stats handler instead.
+func MethodName(n string) otelgrpc.InterceptorFilter {
+	return func(i *otelgrpc.InterceptorInfo) bool {
 		p := splitFullMethod(i)
 		return p.method == n
 	}
@@ -84,8 +104,10 @@ func MethodName(n string) otelgrpc.Filter {
 
 // MethodPrefix returns a Filter that returns true if the request's
 // method starts with the provided string pre.
-func MethodPrefix(pre string) otelgrpc.Filter {
-	return func(i *stats.RPCTagInfo) bool {
+//
+// Deprecated: Use stats handler instead.
+func MethodPrefix(pre string) otelgrpc.InterceptorFilter {
+	return func(i *otelgrpc.InterceptorInfo) bool {
 		p := splitFullMethod(i)
 		return strings.HasPrefix(p.method, pre)
 	}
@@ -94,16 +116,31 @@ func MethodPrefix(pre string) otelgrpc.Filter {
 // FullMethodName returns a Filter that returns true if the request's
 // full RPC method string, i.e. /package.service/method, starts with
 // the provided string n.
-func FullMethodName(n string) otelgrpc.Filter {
-	return func(i *stats.RPCTagInfo) bool {
-		return i.FullMethodName == n
+//
+// Deprecated: Use stats handler instead.
+func FullMethodName(n string) otelgrpc.InterceptorFilter {
+	return func(i *otelgrpc.InterceptorInfo) bool {
+		var fm string
+		switch i.Type {
+		case otelgrpc.UnaryClient, otelgrpc.StreamClient:
+			fm = i.Method
+		case otelgrpc.UnaryServer:
+			fm = i.UnaryServerInfo.FullMethod
+		case otelgrpc.StreamServer:
+			fm = i.StreamServerInfo.FullMethod
+		default:
+			fm = i.Method
+		}
+		return fm == n
 	}
 }
 
 // ServiceName returns a Filter that returns true if the request's
 // service name, i.e. package.service, matches s.
-func ServiceName(s string) otelgrpc.Filter {
-	return func(i *stats.RPCTagInfo) bool {
+//
+// Deprecated: Use stats handler instead.
+func ServiceName(s string) otelgrpc.InterceptorFilter {
+	return func(i *otelgrpc.InterceptorInfo) bool {
 		p := splitFullMethod(i)
 		return p.service == s
 	}
@@ -111,8 +148,10 @@ func ServiceName(s string) otelgrpc.Filter {
 
 // ServicePrefix returns a Filter that returns true if the request's
 // service name, i.e. package.service, starts with the provided string pre.
-func ServicePrefix(pre string) otelgrpc.Filter {
-	return func(i *stats.RPCTagInfo) bool {
+//
+// Deprecated: Use stats handler instead.
+func ServicePrefix(pre string) otelgrpc.InterceptorFilter {
+	return func(i *otelgrpc.InterceptorInfo) bool {
 		p := splitFullMethod(i)
 		return strings.HasPrefix(p.service, pre)
 	}
@@ -121,6 +160,8 @@ func ServicePrefix(pre string) otelgrpc.Filter {
 // HealthCheck returns a Filter that returns true if the request's
 // service name is health check defined by gRPC Health Checking Protocol.
 // https://github.com/grpc/grpc/blob/master/doc/health-checking.md
-func HealthCheck() otelgrpc.Filter {
+//
+// Deprecated: Use stats handler instead.
+func HealthCheck() otelgrpc.InterceptorFilter {
 	return ServicePrefix("grpc.health.v1.Health")
 }